ElementsWidget: Move Delegate to implementation
===============================================

Only moving code from header to implementation.
