+++ src/Mod/Sketcher/Gui/TaskSketcherElements.h
@@ -55,27 +55,7 @@ enum class SubElementType {
     none
 };
 
-class ElementItemDelegate : public QStyledItemDelegate
-{
-    Q_OBJECT
-public:
-    explicit ElementItemDelegate(ElementView* parent);
-    ~ElementItemDelegate() override;
 
-    void paint(QPainter* painter, const QStyleOptionViewItem& option, const QModelIndex& index) const override;
-    bool editorEvent(QEvent* event, QAbstractItemModel* model, const QStyleOptionViewItem& option, const QModelIndex& index) override;
-
-    ElementItem* getElementtItem(const QModelIndex& index) const;
-
-    const int border = 1; //1px, looks good around buttons.
-    const int leftMargin = 4; //4px on the left of icons, looks good.
-    mutable int customIconsMargin = 4;
-    const int textBottomMargin = 5; //5px center the text.
-
-Q_SIGNALS:
-    void itemHovered(QModelIndex);
-    void itemChecked(QModelIndex, Qt::CheckState state);
-};
 
 class ElementView : public QListWidget
 {
